[java] ちょっと ObjectConstrucotr でリファクタ、 input Body を println
